Got Parallel stuff to work

+++ agtlib/cooperative/ppo.py
@@ -241,8 +241,7 @@ class PPO(RLBase):
 
                 entropy_losses.append(entropy_loss.item())
 
-                loss = policy_loss + self.ent_coef * entropy_loss + self.vf_coef * value_loss
-
+                loss = (policy_loss.to(torch.float32) + self.ent_coef * entropy_loss + self.vf_coef * value_loss.to(torch.float32)).float()
                 # Calculate approximate form of reverse KL Divergence for early stopping
                 # see issue #417: https://github.com/DLR-RM/stable-baselines3/issues/417
                 # and discussion in PR #419: https://github.com/DLR-RM/stable-baselines3/pull/419
